code/requirements.txt
pandas
numpy
mysql-connector-python
python-dotenv
google-cloud-storage
